Fixed some renaming issues.
